boolbvt: introduce scopes
This is the biggest change to the solver interface since it got added to the
codebase.

Currently, bound identifiers (exists, forall, let) must be unique; the
solver looses soundness if that assumption is violated.

This change introduces scopes, i.e., the uniqueness requirement is dropped.
Nested scoping works as expected.  This matches what SMT-LIB does, i.e., the
renaming done in the SMT-LIB front-end can be dropped (which is the next
commit).